Create  breadth-first
A breadth-first algorithm (BFS) is a graph traversal or search algorithm that explores the graph level by level, visiting all neighbors of a node before moving on to the neighbors of those neighbors. It begins at the root (or an arbitrary node) and explores all neighboring nodes at the present depth level before moving on to nodes at the next depth level.

Key Properties of BFS:
Level-order traversal: BFS visits all nodes at distance n before visiting nodes at distance n+1.
Queue-based: BFS uses a queue to keep track of the nodes to be visited. The queue follows a First-In-First-Out (FIFO) structure, meaning the first node to be enqueued is the first one to be dequeued.
Shortest Path: BFS can be used to find the shortest path in an unweighted graph.
BFS Algorithm Steps:
Start from a given node (root or any arbitrary node).
Visit the node and mark it as visited.
Add all its unvisited neighbors to a queue.
Dequeue a node from the queue, visit it, and repeat the process until the queue is empty (i.e., all nodes have been visited).